moved uresume routes to uresume controller
