+++ server.js
@@ -92,95 +92,6 @@ function isLoggedIn(req, res, next) {
 
 
 // get and post routes for submitted a new resume - belong in UserResume controller eventually:
-app.get('/newresume', isLoggedIn, function (req, res) {
-   res.render('new_resume', {req:req})
-})
-
-app.post('/newresume', function (req,res) {
-  //sub == 'submission'
-  var sub= req.body
-  var newResume = new UserResume();
-  //make a new resume, give all submission data to its appropriate keys
-  newResume.name = sub.Name;
-  newResume.streetAddress = sub.Address;
-  newResume.email = sub.Email;
-  newResume.phone = sub.Phone;
-  newResume.education = sub.Education
-  newResume.summary = sub.Summary
-
-  //create a new experience object for the array of experience for each new experience that gets added
-  newResume.experiences = [];
-  var experienceObject = {}
-  sub.textinput.forEach(function (input, index) {
-
-    switch(index%4){
-      case 0:
-      experienceObject.title = input
-      break;
-      case 1:
-      experienceObject.startDate = input;
-      break;
-      case 2:
-      experienceObject.endDate = input;
-      break;
-      case 3:
-      experienceObject.description = input;
-      newResume.experiences.push(experienceObject)
-      break;
-    }
-  })
-
-  //give new resume to the current user
-  req.user.resumes.push(newResume);
-  //save unpdated current user and the resume itself into the db
-  newResume.save(function(err) {
-      if (err)
-          throw err;
-  });
-  req.user.save(function(err) {
-      if (err)
-          throw err;
-  });
-
-  res.render('profile', {req: req, user : req.user})
-})
-
-
-app.get('/resume/edit/:id', function (req, res) {
-//  var UserResume.findById(req.params.id);
-
-
-});
-
-app.get('/resume/delete/:id', function (req, res) {
-  var ID = req.params.id
-  UserResume.remove({_id: ID}, function (err, result) {
-            if (err) {
-              console.log(err);
-            }else {
-              console.log('deleted');
-              req.user.resumes=[];
-              req.user.save;
-              // resumes.forEach(function (resume, index) {
-              //   if (resume._id==ID) {
-              //     console.log('found ya');
-              //     resumes.splice(index,1)
-              //     req.user.save(function (err) {
-              //       if (err) {
-              //         throw err
-              //       }
-              //     })
-              //   }
-              // })
-            }
-        })
-
-
-
-
-      res.render('profile', {req: req, user : req.user})
-}
-)
 
 
 fs.readdirSync('./controllers').forEach(function (file) {
